requirements.txt
python-dateutil
django-tastypie
django-widget-tweaks
hexdump
pymongo
python-magic
pytz